made the project support emscripten

+++ Loader.cpp
@@ -252,7 +252,7 @@ void Loader::writeMipMaps(Device device, Texture texture, Extent3D textureSize, 
 void Loader::populateTextureFrameAttributes(std::vector<VertexAttributes>& vertexData) {
 	size_t triangleCount = vertexData.size() / 3;
 	// We compute the local texture frame per triangle
-	for (int t = 0; t < triangleCount; ++t) {
+	for (int t = 0; t < (int)triangleCount; ++t) {
 		VertexAttributes* v = &vertexData[3 * t];
 
 		for (int k = 0; k < 3; ++k) {
